BDD variable and BDD branch node classes
